finished sg_splits
