+++ contracts/splits/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod contract;
 mod error;
 pub mod helpers;
-pub mod integration_tests;
 pub mod msg;
 pub mod state;
