+++ test-suite/src/common_setup.rs
@@ -1,4 +1,6 @@
 pub mod contract_boxes;
+pub mod contract_boxes_empty;
+pub mod helpers;
 pub mod msg;
 pub mod setup_accounts_and_block;
 pub mod setup_base_minter;
