+++ test-suite/src/splits.rs
@@ -1,0 +1,1 @@
+mod tests;
