+++ test-suite/src/splits/tests.rs
@@ -1,0 +1,1 @@
+mod integration_tests;
